don't run the ui locally when running tests
