+++ package.json
@@ -43,13 +43,12 @@
     "html-webpack-plugin": "3.2.0",
     "jest": "23.6.0",
     "react-test-renderer": "16.6.3",
-    "testcafe": "1.2.0",
-    "testcafe-react-selectors": "0.0.8-alpha3",
     "transfer-webpack-plugin": "0.1.4",
     "uglifyjs-webpack-plugin": "v2.0.1",
     "webpack-dev-server": "3.1.10"
   },
   "dependencies": {
+    "@auth0/s3": "^1.0.0",
     "@material-ui/codemod": "^1.1.0",
     "@material-ui/core": "^3.5.1",
     "@material-ui/docs": "^3.0.0-alpha.7",
@@ -80,7 +79,9 @@
     "recharts": "1.3.6",
     "recompose": "^0.30.0",
     "request": "2.88.0",
+    "testcafe": "1.2.0",
     "testcafe-browser-provider-selenium": "^0.4.0",
+    "testcafe-react-selectors": "0.0.8-alpha3",
     "transfer-webpack-plugin": "0.1.4",
     "webpack": "^4.25.1",
     "webpack-cli": "^3.1.2",
